feat: split node format
